requirements.txt
asgiref==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-allauth
Pillow
Django>=3.1.9
psycopg2
psycopg2-binary
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
